Update and rename Token.sol to SimpleToken.sol
style nits (`uint256`)

+++ contracts/token/erc20/SimpleToken.sol
@@ -5,15 +5,15 @@ contract Token {
     string public name;
     string public symbol;
     uint8 constant public decimals = 18;
-    uint public totalSupply;
+    uint256 public totalSupply;
     
-    mapping(address => uint) public balanceOf;
-    mapping(address => mapping(address => uint)) public allowance;
+    mapping(address => uint256) public balanceOf;
+    mapping(address => mapping(address => uint256)) public allowance;
     
-    event Transfer(address indexed from, address indexed to, uint amount);
-    event Approval(address indexed owner, address indexed spender, uint amount);
+    event Transfer(address indexed from, address indexed to, uint256 amount);
+    event Approval(address indexed owner, address indexed spender, uint256 amount);
     
-    constructor(address owner, string memory _name, string memory _symbol, uint _totalSupply) {
+    constructor(address owner, string memory _name, string memory _symbol, uint256 _totalSupply) {
         name = _name;
         symbol = _symbol;
         totalSupply = _totalSupply;
@@ -21,21 +21,21 @@ contract Token {
         emit Transfer(address(0), owner, _totalSupply);
     }
     
-    function approve(address to, uint amount) external returns (bool) {
+    function approve(address to, uint256 amount) external returns (bool) {
         allowance[msg.sender][to] = amount;
         emit Approval(msg.sender, to, amount);
         return true;
     }
     
-    function transfer(address to, uint amount) external returns (bool) {
+    function transfer(address to, uint256 amount) external returns (bool) {
         balanceOf[msg.sender] -= amount;
         balanceOf[to] += amount;
         emit Transfer(msg.sender, to, amount);
         return true;
     }
     
-    function transferFrom(address from, address to, uint amount) external returns (bool) {
-        if (allowance[from][msg.sender] != type(uint).max) 
+    function transferFrom(address from, address to, uint256 amount) external returns (bool) {
+        if (allowance[from][msg.sender] != type(uint256).max) 
             allowance[from][msg.sender] -= amount;
         balanceOf[from] -= amount;
         balanceOf[to] += amount;
